Tighten up warnings: * Remove several warning suppression flags from CMakeLists.txt
* Add three new targeted warning suppression macros in config.hpp:
  * RANGES_DIAGNOSTIC_IGNORE_GLOBAL_CONSTRUCTORS
  * RANGES_DIAGNOSTIC_IGNORE_SIGN_CONVERSION
  * RANGES_DIAGNOSTIC_IGNORE_UNNEEDED_INTERNAL
  These are used to suppress targeted warnings in the tests.

* By default, RANGES_DISABLE_WARNINGS and RANGES_RE_ENABLE_WARNINGS are now no-ops. We want to (a) use the test suite to catch questionable usage in library headers, (b) expose diagnostics to users who use the library in questionable ways. Disabling a large class of warnings in the library headers impedes both of those goals. This is a breaking change for code compiling with warnings-as-errors that may now diagnose errors that were suppressed before. Users who just don't care and want the old behavior can define RANGES_FEWER_WARNINGS to get the old behavior.

* Suppress a couple of dozen sign-conversion and 64-to-32 truncation warnings. Many remain as exercises for the reader.

* In functional.hpp, implement class template convert_to<T>, a function object that returns its argument converted to T.

* Relocate the code of some of the tests into an unnamed namespace since it does not need external linkage. This is mostly a style preference.

+++ test/to_container.cpp
@@ -51,7 +51,7 @@ int main()
     ::check_equal(vec0, {81,64,49,36,25,16,9,4,1,0});
 
     auto vec1 = view::ints | view::transform([](int i){return i*i;}) | view::take(10)
-        | to_<std::vector<long>>() | action::sort(std::greater<int>{});
+        | to_<std::vector<long>>() | action::sort(std::greater<long>{});
     static_assert((bool)Same<decltype(vec1), std::vector<long>>(), "");
     ::check_equal(vec1, {81,64,49,36,25,16,9,4,1,0});
